+++ run/inc/dev/vscodium.sh
@@ -1,0 +1,14 @@
+#!/usr/bin/env bash
+
+#
+# Install VS Codium – a fully free/libre version of Visual Studio Code.
+#
+# https://github.com/VSCodium/vscodium/
+#
+
+is_gui_enabled || return 0
+
+print_step "Installing VS Codium."
+
+print_info "Installing/updating VS Codium via APT."
+superdo apt-get install -y codium